Fix center and zoom
